Refactor internal signature creation into multiple files
