+++ src/Phan/Language/Internal/FunctionSignatureMap.php
@@ -7162,7 +7162,7 @@ return [
 'MongoDB\BSON\regexinterface::getFlags' => ['string'],
 'MongoDB\BSON\regexinterface::getPattern' => ['string'],
 'MongoDB\BSON\Serializable::bsonSerialize' => ['array|object'],
-'MongoDB\BSON\symbol::__construct' => [''],
+'MongoDB\BSON\symbol::__construct' => ['void'],
 'MongoDB\BSON\symbol::__toString' => ['string'],
 'MongoDB\BSON\symbol::jsonSerialize' => ['mixed'],
 'MongoDB\BSON\symbol::serialize' => ['string'],
@@ -7179,7 +7179,7 @@ return [
 'MongoDB\BSON\timestampinterface::getTimestamp' => ['int'],
 'MongoDB\BSON\toJSON' => ['string', 'bson'=>'string'],
 'MongoDB\BSON\toPHP' => ['object', 'bson'=>'string', 'typeMap'=>'array'],
-'MongoDB\BSON\undefined::__construct' => [''],
+'MongoDB\BSON\undefined::__construct' => ['void'],
 'MongoDB\BSON\undefined::__toString' => ['string'],
 'MongoDB\BSON\undefined::jsonSerialize' => ['mixed'],
 'MongoDB\BSON\undefined::serialize' => ['string'],
@@ -7300,7 +7300,7 @@ return [
 'MongoDB\Driver\Server::isPassive' => ['bool'],
 'MongoDB\Driver\Server::isPrimary' => ['bool'],
 'MongoDB\Driver\Server::isSecondary' => ['bool'],
-'mongodb\driver\session::__construct' => [''],
+'mongodb\driver\session::__construct' => ['void'],
 'mongodb\driver\session::abortTransaction' => ['void'],
 'mongodb\driver\session::advanceClusterTime' => ['void', 'clusterTime'=>'array|object'],
 'mongodb\driver\session::advanceOperationTime' => ['void', 'operationTime'=>'MongoDB\BSON\TimestampInterface'],
